requirements.txt
streamlit 
python-docx 
transformers 
sentence-transformers 
difflib